+++ CHANGELOG.md
@@ -2,6 +2,44 @@
 
 ## [Unreleased][unreleased]
 
+### Added
+### Changed
+### Removed
+### Fixed
+### Deprecated
+
+## [v0.6.1][]
+
+This minor release fixes a number of bugs with the ACE interface and a
+number of small bugs that came up during testing. While a minor release,
+one new (minor) feature is added: an AceTransferer and associated
+functions.
+
+### Added
+
+* `delphin.interfaces.ace.AceTransferer` - due to a limitation with ACE,
+  only the regular (non-tsdb) mode is allowed in transfer
+* `delphin.interfaces.ace.transfer()`
+* `delphin.interfaces.ace.transfer_from_iterable()`
+
+### Changed
+
+* `delphin.interfaces.ace` no longer captures stderr
+* `delphin.interfaces.ace` attempts to restart a process that closed
+  unexpectedly
+
+### Fixed
+
+* `delphin.interfaces.ace` ignores stderr, which effectively fixes #86
+* `delphin.interfaces.ace` joins content lines in tsdb mode---fixes #95
+* Write empty tables for `delphin mkprof` command unless `--skeleton` is
+  used (fixes #96)
+* `delphin.mrs.xmrs.Xmrs` equality test now checks variable properties
+* `delphin.mrs.xmrs.Mrs.from_dict()` correctly reads variable properties (#98)
+* `delphin.interfaces.ace._AceResponse.get()` now backs off to ParseResponse
+  on KeyError
+* `delphin.mrs.simplemrs` fix error message on unexpected token
+
 ## [v0.6.0][]
 
 This release replaces the top-level `pyDelphin` and `mrs.py` scripts
@@ -414,6 +452,7 @@ information about changes, except for
 [commit messages](../../commits/v0.2).
 
 [unreleased]: ../../tree/develop
+[v0.6.1]: ../../releases/tag/v0.6.1
 [v0.6.0]: ../../releases/tag/v0.6.0
 [v0.5.1]: ../../releases/tag/v0.5.1
 [v0.5.0]: ../../releases/tag/v0.5.0